SoCs for MVT
