+++ src/Components/Pages/Dashboard/Graphs/MVT/Graph.js
@@ -1,0 +1,83 @@
+import { ResponsiveLine } from "@nivo/line";
+
+const MvtGraph = ({ points }) => {
+    return (
+        <ResponsiveLine
+            points={points}
+            // TODO: Uncomment the below code and remove the above data passing code 
+            // data={points}
+            margin={{ top: 50, right: 110, bottom: 50, left: 60 }}
+            xScale={{
+                type: "linear",
+                min: "auto",
+                max: "auto",
+                stacked: false,
+                reverse: false,
+            }}
+            yScale={{
+                type: "linear",
+                min: "auto",
+                max: "auto",
+                stacked: false,
+                reverse: false,
+            }}
+            xFormat=' >-.2f'
+            yFormat=' >-.2f'
+            axisTop={null}
+            axisRight={null}
+            axisBottom={{
+                tickSize: 5,
+                tickPadding: 5,
+                tickRotation: 0,
+                legend: "Time Since Replica Started Accepting Messages(10^-5 seconds)",
+                legendOffset: 36,
+                legendPosition: "middle",
+                fontColor: '#fff'
+            }}
+            axisLeft={{
+                tickSize: 5,
+                tickPadding: 5,
+                tickRotation: 0,
+                legend: "Number of Messages",
+                legendOffset: -40,
+                legendPosition: "middle",
+            }}
+            enablePoints={true}
+            pointSize={10}
+            pointColor={{ theme: "background" }}
+            pointBorderWidth={2}
+            pointBorderColor={{ from: "serieColor" }}
+            pointLabelYOffset={- 12}
+            useMesh={true}
+            legends={
+                [
+                    {
+                        anchor: "bottom-right",
+                        direction: "column",
+                        justify: false,
+                        translateX: 100,
+                        translateY: 0,
+                        itemsSpacing: 0,
+                        itemDirection: "left-to-right",
+                        itemWidth: 80,
+                        itemHeight: 20,
+                        itemOpacity: 0.75,
+                        symbolSize: 12,
+                        symbolShape: "circle",
+                        symbolBorderColor: "rgba(0, 0, 0, .5)",
+                        effects: [
+                            {
+                                on: "hover",
+                                style: {
+                                    itemBackground: "rgba(0, 0, 0, .03)",
+                                    itemOpacity: 1,
+                                },
+                            },
+                        ],
+                    },
+                ]}
+        />
+    );
+}
+
+export default MvtGraph
